+++ options/options_test.cc
@@ -3052,6 +3052,15 @@ TEST_F(OptionsParserTest, DifferentDefault) {
     old_default_opts.OldDefaults(5, 18);
     ASSERT_TRUE(old_default_opts.compaction_pri ==
                 CompactionPri::kByCompensatedSize);
+    ASSERT_EQ(0, old_default_opts.memtable_prefix_bloom_size_ratio);
+    ASSERT_FALSE(old_default_opts.memtable_whole_key_filtering);
+  }
+
+  {
+    Options old_default_opts;
+    old_default_opts.OldDefaults(6, 17);
+    ASSERT_EQ(0, old_default_opts.memtable_prefix_bloom_size_ratio);
+    ASSERT_FALSE(old_default_opts.memtable_whole_key_filtering);
   }
 
   Options small_opts;